+++ contract-testing/src/invocation/types.rs
@@ -1,5 +1,5 @@
 use crate::Chain;
-use concordium_base::{
+use concordium_rust_sdk::base::{
     base::{AccountAddressEq, Energy},
     contracts_common::{
         AccountAddress, Address, Amount, ContractAddress, ModuleReference, OwnedContractName,
@@ -8,10 +8,10 @@ use concordium_base::{
     smart_contracts::OwnedParameter,
     transactions::UpdateContractPayload,
 };
-use concordium_smart_contract_engine::v1::{
+use concordium_rust_sdk::smart_contracts::engine::v1::{
     trie::MutableState, DebugTracker, InvokeResponse, ReceiveContext, ReceiveInterruptedState,
 };
-use concordium_wasm::artifact::CompiledFunction;
+use concordium_rust_sdk::smart_contracts::engine::wasm::artifact::CompiledFunction;
 use std::collections::BTreeMap;
 
 /// A type that supports invoking a contract entrypoint.