+++ contract-testing/tests/transfer.rs
@@ -110,7 +110,7 @@ fn test_transfer() {
                 address:          contract_address,
                 amount:           Amount::zero(),
                 receive_name:     OwnedReceiveName::new_unchecked("transfer.send".into()),
-                contract_version: concordium_base::smart_contracts::WasmVersion::V1,
+                contract_version: WasmVersion::V1,
                 instigator:       Address::Account(helpers::ACC_0),
                 message:          parameter,
                 events:           Vec::new(),